Renames generic "prompts" to "add-prompts"
Renames generic "prompts" to "add-prompts"
